abstracted useVisualisation hook into its own file
